+++ libtransport/src/implementation/tls_socket_consumer.cc
@@ -75,7 +75,7 @@ int readOldTLS(BIO *b, char *buf, int size) {
     }
   }
 
-  return read;
+  return (int)read;
 }
 
 /* Return the number of read bytes in readbytes */
